// src/app/shared/services/model/model.ts
export interface apiResultFormat {
  data: [];
  totalData: number;
}
// ...
export interface header {
  tittle: string;
  showAsTab: boolean;
  separateRoute: boolean;
  routes?: string;
  hasSubRoute?: boolean;
  showSubRoute?: boolean;
  menu?: menu[];
  base?: string;
}
export interface menu {
  menuValue: string;
  routes?: string;
  hasSubRoute: boolean;
  showSubRoute: boolean;
  subMenus?: submenu[];
  page1?: string;
  page2?: string;
  page3?: string;
  page4?: string;
  page5?: string;
  page6?: string;
  page7?: string;
  page8?: string;
  page9?: string;
  page10?: string;
  page11?: string;
}
export interface submenu {
  menuValue: string;
  routes: string;
  hasSubRoute: boolean;
  showSubRoute: boolean;
  subMenus: [];
}